backend/requirements.txt
# Django Core
django>=5.2
django-ninja>=1.3.0
django-ninja-jwt>=5.3.0
django-ninja-extra>=0.30.6  # django-ninja-jwt 的依赖

# Database & Cache
django-redis>=6.0.0

# Background Tasks
django-q2>=1.6.0

# Data Processing
pydantic>=2.0
pandas>=2.0
openpyxl>=3.1.0

# HTTP Client
httpx>=0.27.0  # Async HTTP client for insurance API
requests>=2.31.0  # Sync HTTP client

# WSGI Server
gunicorn>=21.0.0  # Production WSGI server

# Document Processing
playwright>=1.40.0
docxtpl>=0.16.0
pymupdf>=1.24.0  # PDF processing (fitz module)
python-docx>=1.1.0  # Word document processing
rapidocr-onnxruntime>=1.3.0  # OCR for document text extraction

# OCR for Captcha
ddddocr>=1.4.0

# Utilities
python-dotenv>=1.0.0
watchdog>=3.0.0  # File system monitoring for config hot reload
pyyaml>=6.0.0  # YAML configuration support

# CORS & Security
django-cors-headers>=4.3.0
cryptography>=41.0.0  # 爬虫配置加密

# Admin
django-nested-admin>=4.0.0

# Development & Testing
pytest>=8.0.0
pytest-django>=4.5.0
pytest-asyncio>=0.23.0  # Async test support
pytest-cov>=4.0.0
factory-boy>=3.3.0
hypothesis>=6.100.0  # Property-based testing

# Code Quality (optional)
# black>=24.0.0
# isort>=5.13.0
# flake8>=7.0.0
# mypy>=1.8.0
